@@ -5,12 +5,13 @@ import com.velocitypowered.api.plugin.annotation.DataDirectory;
 import com.velocitypowered.api.proxy.ProxyServer;
 import dev.scarday.command.HubCommand;
 import dev.scarday.config.Configuration;
+import dev.scarday.hub.factory.ServerFactory;
+import dev.scarday.hub.factory.ServerFactoryImpl;
 import eu.okaeri.configs.ConfigManager;
 import eu.okaeri.configs.serdes.adventure.SerdesAdventure;
 import eu.okaeri.configs.yaml.snakeyaml.YamlSnakeYamlConfigurer;
 import lombok.Getter;
 import lombok.val;
-import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import javax.inject.Inject;
 import com.velocitypowered.api.event.Subscribe;
@@ -19,32 +20,33 @@ import com.velocitypowered.api.event.proxy.ProxyInitializeEvent;
 import java.io.File;
 import java.nio.file.Path;
 
-@Plugin(id = "hub", name = "Hub", version = BuildConstant.VERSION, authors = "ScarDay")
 @Getter
+@Plugin(id = "hub", name = "Hub", version = "1.0", authors = "ScarDay")
 public class Main {
     final ProxyServer proxy;
     final Logger logger;
     File dataDirectory;
 
-    @NotNull
-    Configuration config = new Configuration();
+    Configuration config;
+
+    ServerFactory serverFactory;
 
     @Inject
     public Main(ProxyServer server, Logger logger, @DataDirectory Path dataDirectory) {
         this.proxy = server;
         this.logger = logger;
 
-        this.dataDirectory = dataDirectory
-                .toFile();
+        this.dataDirectory = dataDirectory.toFile();
+        this.serverFactory = new ServerFactoryImpl(this);
     }
 
     @Subscribe
     public void onProxyInitialization(ProxyInitializeEvent event) {
         initConfig();
-        registerCommands();
+        registerCommand();
     }
 
-    private void registerCommands() {
+    private void registerCommand() {
         val commandManager = getProxy().getCommandManager();
         val command = new HubCommand(this);
 
@@ -59,8 +61,6 @@ public class Main {
                 it.saveDefaults();
                 it.load(true);
             });
-
-            logger.info("Configuration 'config.yml' is loaded!");
         } catch (Exception exception) {
             getLogger().error("Error loading config.yml", exception);
         }
