@@ -1,4 +1,4 @@
-package dev.scarday.multihub;
+package dev.scarday.hub.type;
 
 public enum Type {
     NONE,
